Refactors API calls and adds a call to get a list of events
