+++ src/DevSoc.Eventathon/ClientApp/src/api/health-endpoint.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import useSWR from 'swr';
 
-export function useHealthCheck () {
+export const checkHealth = () => {
     const fetcher = url => axios.get(url).then(response => response.status);
     return useSWR('api/health', fetcher);
   }
